release 1.10.0

+++ bot/helpers/handlers/handle_replicate_webhook.py
@@ -18,51 +18,64 @@ from bot.database.operations.request.getters import get_request
 from bot.database.operations.request.updaters import update_request
 from bot.database.operations.transaction.writers import write_transaction
 from bot.database.operations.user.getters import get_user
-from bot.database.operations.user.updaters import update_user
 from bot.handlers.ai.face_swap_handler import PRICE_FACE_SWAP, handle_face_swap
 from bot.handlers.ai.music_gen_handler import PRICE_MUSIC_GEN, handle_music_gen
+from bot.handlers.ai.stable_diffusion_handler import PRICE_STABLE_DIFFUSION
 from bot.helpers.senders.send_audio import send_audio
 from bot.helpers.senders.send_images import send_image
-from bot.keyboards.common.common import build_reaction_keyboard
+from bot.helpers.updaters.update_user_usage_quota import update_user_usage_quota
+from bot.keyboards.common.common import build_reaction_keyboard, build_error_keyboard
 from bot.locales.main import get_user_language, get_localization
 
 
 async def handle_replicate_webhook(bot: Bot, dp: Dispatcher, prediction: dict):
-    generation = await get_generation(prediction.get("id"))
+    generation = await get_generation(prediction.get('id'))
     if not generation:
         return False
     elif generation.status == GenerationStatus.FINISHED:
         return True
 
-    generation_error, generation_result = prediction.get("error", False), prediction.get("output", {})
-    seconds = prediction.get("metrics", {}).get("predict_time", 0)
+    request = await get_request(generation.request_id)
+    user = await get_user(request.user_id)
+
+    generation_error, generation_result = prediction.get('error', False), prediction.get('output', {})
+    seconds = prediction.get('metrics', {}).get('predict_time', 0)
 
     generation.status = GenerationStatus.FINISHED
     generation.seconds = seconds
-    if generation_error or not generation_result:
+    if generation_error and 'NSFW content detected' in generation_error:
+        await update_generation(generation.id, {
+            'status': generation.status,
+            'seconds': generation.seconds,
+        })
+
+        await bot.send_message(
+            chat_id=user.telegram_chat_id,
+            text=get_localization(user.interface_language_code).REQUEST_FORBIDDEN_ERROR,
+        )
+    elif generation_error or not generation_result:
         generation.has_error = True
         await update_generation(generation.id, {
-            "status": generation.status,
-            "has_error": generation.has_error,
-            "seconds": generation.seconds,
+            'status': generation.status,
+            'has_error': generation.has_error,
+            'seconds': generation.seconds,
         })
 
-        logging.error(f"Error in replicate_webhook: {prediction.get('logs')}")
+        logging.error(f'Error in replicate_webhook: {prediction.get("logs")}')
     else:
-        generation.result = generation_result
+        generation.result = generation_result[0] if type(generation_result) == list else generation_result
         await update_generation(generation.id, {
-            "status": generation.status,
-            "result": generation.result,
-            "seconds": generation.seconds,
+            'status': generation.status,
+            'result': generation.result,
+            'seconds': generation.seconds,
         })
-
-    request = await get_request(generation.request_id)
-    user = await get_user(request.user_id)
 
     if request.model == Model.FACE_SWAP:
         await handle_replicate_face_swap(bot, dp, user, request, generation)
     elif request.model == Model.MUSIC_GEN:
         await handle_replicate_music_gen(bot, dp, user, request, generation)
+    elif request.model == Model.STABLE_DIFFUSION:
+        await handle_replicate_stable_diffusion(bot, dp, user, request, generation)
 
     return True
 
@@ -82,7 +95,7 @@ async def handle_replicate_face_swap(
     if current_count == request.requested and request.status != RequestStatus.FINISHED:
         request.status = RequestStatus.FINISHED
         await update_request(request.id, {
-            "status": request.status
+            'status': request.status
         })
 
         request_generations = await get_generations_by_request_id(request.id)
@@ -107,9 +120,9 @@ async def handle_replicate_face_swap(
             )
 
         used_face_swap_package = await get_used_face_swap_package(
-            generation.details.get("used_face_swap_package_id")
+            generation.details.get('used_face_swap_package_id')
         )
-        if request.details.get("is_test"):
+        if request.details.get('is_test'):
             total_price = round(PRICE_FACE_SWAP * total_seconds, 6)
             update_tasks = [
                 write_transaction(
@@ -121,25 +134,13 @@ async def handle_replicate_face_swap(
                     currency=Currency.USD,
                     quantity=total_result,
                     details={
-                        'name': request.details.get("face_swap_package_name"),
+                        'name': request.details.get('face_swap_package_name'),
                         'images': success_generations,
                         'seconds': total_seconds,
                         'has_error': generation.has_error,
                     }),
             ]
         else:
-            quantity_to_delete = total_result
-            quantity_deleted = 0
-            while quantity_deleted != quantity_to_delete:
-                if user.monthly_limits[Quota.FACE_SWAP] != 0:
-                    user.monthly_limits[Quota.FACE_SWAP] -= 1
-                    quantity_deleted += 1
-                elif user.additional_usage_quota[Quota.FACE_SWAP] != 0:
-                    user.additional_usage_quota[Quota.FACE_SWAP] -= 1
-                    quantity_deleted += 1
-                else:
-                    break
-
             total_price = round(PRICE_FACE_SWAP * total_seconds, 6)
             update_tasks = [
                 write_transaction(
@@ -149,19 +150,21 @@ async def handle_replicate_face_swap(
                     amount=total_price,
                     clear_amount=total_price,
                     currency=Currency.USD,
-                    quantity=quantity_to_delete,
+                    quantity=total_result,
                     details={
-                        'name': request.details.get("face_swap_package_name", "CUSTOM"),
+                        'name': request.details.get('face_swap_package_name', 'CUSTOM'),
                         'images': success_generations,
                         'seconds': total_seconds,
                         'has_error': generation.has_error,
                     }
                 ),
-                update_user(user.id, {
-                    "monthly_limits": user.monthly_limits,
-                    "additional_usage_quota": user.additional_usage_quota
-                }),
+                update_user_usage_quota(
+                    user,
+                    Quota.FACE_SWAP,
+                    total_result,
+                ),
             ]
+
             if (
                 total_result == len(request_generations) and
                 used_face_swap_package and
@@ -169,7 +172,7 @@ async def handle_replicate_face_swap(
             ):
                 update_tasks.append(
                     update_used_face_swap_package(used_face_swap_package.id, {
-                        "used_images": used_face_swap_package.used_images + used_face_swap_package_used_images,
+                        'used_images': used_face_swap_package.used_images + used_face_swap_package_used_images,
                     })
                 )
 
@@ -203,9 +206,9 @@ async def handle_replicate_music_gen(
 
     if generation.result:
         reply_markup = build_reaction_keyboard(generation.id)
-        caption = f"🎵 {(user.monthly_limits[Quota.MUSIC_GEN] + user.additional_usage_quota[Quota.MUSIC_GEN]) - duration}" \
-            if user.settings[Model.MUSIC_GEN][UserSettings.SHOW_USAGE_QUOTA] \
-            else "🎵"
+        caption = f'🎵 {(user.daily_limits[Quota.MUSIC_GEN] + user.additional_usage_quota[Quota.MUSIC_GEN]) - duration}' \
+            if user.settings[Model.MUSIC_GEN][UserSettings.SHOW_USAGE_QUOTA] and \
+               user.daily_limits[Quota.MUSIC_GEN] != float('inf') else '🎵'
         await send_audio(
             bot,
             user.telegram_chat_id,
@@ -215,24 +218,20 @@ async def handle_replicate_music_gen(
             duration,
             reply_markup,
         )
+    elif generation.has_error:
+        reply_markup = build_error_keyboard(user.interface_language_code)
+        await bot.send_message(
+            chat_id=user.telegram_chat_id,
+            text=get_localization(user.interface_language_code).ERROR,
+            reply_markup=reply_markup,
+            parse_mode=None,
+        )
 
     if request.status != RequestStatus.FINISHED:
         request.status = RequestStatus.FINISHED
         await update_request(request.id, {
-            "status": request.status
+            'status': request.status
         })
-
-        quantity_to_delete = duration if generation.result and duration else 0
-        quantity_deleted = 0
-        while quantity_deleted != quantity_to_delete:
-            if user.monthly_limits[Quota.MUSIC_GEN] != 0:
-                user.monthly_limits[Quota.MUSIC_GEN] -= 1
-                quantity_deleted += 1
-            elif user.additional_usage_quota[Quota.MUSIC_GEN] != 0:
-                user.additional_usage_quota[Quota.MUSIC_GEN] -= 1
-                quantity_deleted += 1
-            else:
-                break
 
         total_price = round(PRICE_MUSIC_GEN * generation.seconds, 6)
         update_tasks = [
@@ -243,7 +242,7 @@ async def handle_replicate_music_gen(
                 amount=total_price,
                 clear_amount=total_price,
                 currency=Currency.USD,
-                quantity=quantity_to_delete,
+                quantity=1 if generation.result else 0,
                 details={
                     'result': generation.result,
                     'prompt': prompt,
@@ -251,10 +250,11 @@ async def handle_replicate_music_gen(
                     'has_error': generation.has_error,
                 },
             ),
-            update_user(user.id, {
-                "monthly_limits": user.monthly_limits,
-                "additional_usage_quota": user.additional_usage_quota
-            }),
+            update_user_usage_quota(
+                user,
+                Quota.MUSIC_GEN,
+                1 if generation.result else 0,
+            ),
         ]
 
         await asyncio.gather(*update_tasks)
@@ -271,5 +271,70 @@ async def handle_replicate_music_gen(
 
         if user.current_model == Model.MUSIC_GEN:
             await handle_music_gen(bot, user.telegram_chat_id, state, user.id)
+
+        await bot.delete_message(user.telegram_chat_id, request.message_id)
+
+
+async def handle_replicate_stable_diffusion(
+    bot: Bot,
+    dp: Dispatcher,
+    user: User,
+    request: Request,
+    generation: Generation,
+):
+    prompt = generation.details.get('prompt')
+
+    if generation.result:
+        reply_markup = build_reaction_keyboard(generation.id)
+        await send_image(bot, user.telegram_chat_id, generation.result, reply_markup)
+    elif generation.has_error:
+        reply_markup = build_error_keyboard(user.interface_language_code)
+        await bot.send_message(
+            chat_id=user.telegram_chat_id,
+            text=get_localization(user.interface_language_code).ERROR,
+            reply_markup=reply_markup,
+            parse_mode=None,
+        )
+
+    if request.status != RequestStatus.FINISHED:
+        request.status = RequestStatus.FINISHED
+        await update_request(request.id, {
+            'status': request.status
+        })
+
+        total_price = round(PRICE_STABLE_DIFFUSION * generation.seconds, 6)
+        update_tasks = [
+            write_transaction(
+                user_id=user.id,
+                type=TransactionType.EXPENSE,
+                service=ServiceType.STABLE_DIFFUSION,
+                amount=total_price,
+                clear_amount=total_price,
+                currency=Currency.USD,
+                quantity=1 if generation.result else 0,
+                details={
+                    'result': generation.result,
+                    'prompt': prompt,
+                    'has_error': generation.has_error,
+                },
+            ),
+            update_user_usage_quota(
+                user,
+                Quota.STABLE_DIFFUSION,
+                1 if generation.result else 0,
+            ),
+        ]
+
+        await asyncio.gather(*update_tasks)
+
+        state = FSMContext(
+            storage=dp.storage,
+            key=StorageKey(
+                chat_id=int(user.telegram_chat_id),
+                user_id=int(user.id),
+                bot_id=bot.id,
+            )
+        )
+        await state.clear()
 
         await bot.delete_message(user.telegram_chat_id, request.message_id)
